completed part 2 of the lab.  Added UML
