+++ src/main/java/Main.java
@@ -12,6 +12,7 @@ public class Main {
 
 
 
+
         petFactory.addPetsToList(petFactory.createPets());
 
         ArrayList<Pet> petList = petFactory.getPetList();
@@ -22,14 +23,25 @@ public class Main {
 
 
         for(Pet pet: petList){
+            System.out.println("Print by Name");
             System.out.println(pet.getPetName());
-        }
-
-        for(Pet pet : petList) {
+            System.out.println(pet.getPetType());
             System.out.println(pet.speak());
+            System.out.println("________________________________");
         }
 
-       
+
+        Collections.sort(petList, new ObjectofPetComparison());
+
+        for(Pet pet1: petList){
+            System.out.println("Print by Type");
+            System.out.println(pet1.getPetType());
+            System.out.println(pet1.getPetName());
+            System.out.println(pet1.speak());
+            System.out.println("_________________________________");
+        }
+
+
